huge update

+++ config/fabric/services/screen_record.py
@@ -11,11 +11,22 @@ class ScreenRecorder(Service):
     @Signal
     def recording(self, value: bool) -> None: ...
 
+    _instance = None
+
     def __init__(self, **kwargs):
+        if hasattr(self, "_initialized") and self._initialized:
+            return
         self.screenshot_path = GLib.get_home_dir() + "/Pictures/Screenshots/"
         self.screenrecord_path = GLib.get_home_dir() + "/Videos/Screencasting/"
+        self._current_screencast_path = None
 
         super().__init__(**kwargs)
+        self._initialized = True
+
+    def __new__(cls, *args, **kwargs):
+        if not cls._instance:
+            cls._instance = super(ScreenRecorder, cls).__new__(cls, *args, **kwargs)
+        return cls._instance
 
     def screenshot(self, fullscreen=False, save_copy=True):
         time = datetime.datetime.today().strftime("%Y-%m-%d_%H-%M-%S")
@@ -44,18 +55,27 @@ class ScreenRecorder(Service):
         time = datetime.datetime.today().strftime("%Y-%m-%d_%H-%M-%S")
         file_path = self.screenrecord_path + str(time) + ".mp4"
         self._current_screencast_path = file_path
-        area = ""
-        if fullscreen:
-            f"-g '{exec_shell_command('slurp')}'"
+        area = "" if fullscreen else f"-g '{exec_shell_command('slurp')}'"
         command = f"wf-recorder --file={file_path} --pixel-format yuv420p {area}"
         exec_shell_command_async(command)
         self.emit("recording", True)
 
     def screencast_stop(self):
+        if not self.is_recording:
+            logger.error("[SCREENRECORD] No active recording to stop")
+            return
         exec_shell_command_async("killall -INT wf-recorder")
         self.emit("recording", False)
         self.send_screencast_notification(self._current_screencast_path)
 
+    # def screencast_stop(self):
+    #     exec_shell_command_async("killall -INT wf-recorder")
+    #     self.emit("recording", False)
+    #     if hasattr(self, "_current_screencast_path") and self._current_screencast_path:
+    #         self.send_screencast_notification(self._current_screencast_path)
+    #     else:
+    #         logger.error("[SCREENRECORD] No active screencast to stop")
+    #
     def send_screencast_notification(self, file_path):
         # TODO: Generate a thumbnail
         # exec_shell_command_async(f"totem-video-thumbnailer {file_path} {FABRIC_CACHE}")
